\name{akmeans-package}
\alias{akmeans-package}
\docType{package}
\title{
Adaptive K-means
}
\description{
This package provides two things.
At first, K-means based on cosine distance is provided as a component of this package. Then, based on existing 'kmeans' function and the new kmeans with cosine distance, adaptive part is implemented using various thresholds.
}
\details{
\tabular{ll}{
Package: \tab akmeans\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-12-18\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
\itemize{
  \item akmeans - the main function, adaptive kmeans algorithm
  \item norm.sim.ksc - kmeans based on cosine distance
}
}
\author{
Jungsuk Kwac <kwjusu1@stanford.edu>
}
