% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akmeans.clust.R
\name{akmeans.clust}
\alias{akmeans.clust}
\title{Clustering of longitudinal data}
\usage{
akmeans.clust(traj, id_field = FALSE, init_method = "lpm", n_clusters = 3)
}
\arguments{
\item{traj}{A matrix or data.frame with each row representing the trajectory of observations of a unique location. The columns show the observations at consecutive time steps.}

\item{id_field}{Whether the first column is a unique (\code{id}) field. Default: \code{FALSE}}

\item{init_method}{initialisation method. Specifying a method to determine the initial centroids for clustering. Default: \code{"lpm"} - linear partitioning medoids @seealso \code{\link{lpm.centroids}}]}

\item{n_clusters}{number of clusters to generate. Default: \code{3}: (minimum value)}
}
\value{
The original (\code{traj}) data with cluster label appended
}
\description{
This function group trajectories based on a given list of initial centroids
}
\details{
Given a list of trajectories represented in a matrix or data.frame, and a method for choosing initial cluster centroids (e.g. \code{\link{lpm.centroids}}), a list of clusters is generated after a limited number of iterations.
}
\examples{
traj <- gm.crime.sample1
print(traj)
traj <- missingValue(traj, id_field = TRUE, method = 2, replace_with = 1,
fill_zeros = FALSE) #filling the missing values
print(traj)
clusters <- akmeans.clust(traj, id_field = TRUE, init_method = "lpm", n_clusters = 3)
print(clusters)
}
\references{
\code{Christophe Genolini, Xavier Alacoque, Marianne Sentenac, Catherine Arnaud (2015). kml and kml3d: R Packages to Cluster Longitudinal Data. Journal of Statistical Software, 65(4), 1-34. URL http://www.jstatsoft.org/v65/i04/}
}
