% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphaLabel.R
\name{alphaLabel}
\alias{alphaLabel}
\title{Function to assign alphabetic labels to numeric cluster IDs}
\usage{
alphaLabel(x = clusters)
}
\arguments{
\item{x}{A vector of numeric cluster ids}
}
\description{
GGGGGGGGGGG
}
