% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assault_data.R
\docType{data}
\name{assault_data}
\alias{assault_data}
\title{Sample crime (assault) dataset}
\format{A matrix}
\usage{
assault_data
}
\description{
Simulated crime dataset with missing values.
}
\keyword{datasets}
