% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caliHarab.criterion.R
\name{caliHarab.criterion}
\alias{caliHarab.criterion}
\title{Calinski and Harabatz quality criterion}
\usage{
caliHarab.criterion(clustr, id_field = FALSE)
}
\arguments{
\item{clustr}{A data.frame of cluster trajectories (e.g. from \code{\link{akmeans.clust}}), in which the last column represents alphabetical cluster ids (labels)}

\item{id_field}{Whether the first column is a unique (id) field. Default: \code{TRUE}.}
}
\description{
Function to calculate the Calinski-Harabatz score for a given clustering solution
}
\references{
\code{Calinski T, Harabasz J (1974) A dendrite method for cluster analysis. Commun Stat 3(1):1-27}
}
