% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm.crime.sample2.R
\docType{data}
\name{gm.crime.sample2}
\alias{gm.crime.sample2}
\title{Sample crime dataset}
\format{A matrix}
\usage{
gm.crime.sample2
}
\description{
Crime dataset of greater Manchester crime data aggregated at the LSOA geographical level data (Source: data.police.uk)
}
\keyword{datasets}
