% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_crime_data.R
\docType{data}
\name{gm_crime_data}
\alias{gm_crime_data}
\title{Sample crime dataset}
\format{A matrix}
\usage{
gm_crime_data
}
\description{
Crime dataset of greater Manchester crime data aggregated at the LSOA geographical level data (Source: data.police.uk)
}
\keyword{datasets}
