% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpm.centroids.R
\name{lpm.centroids}
\alias{lpm.centroids}
\title{Linear Partition Medoids (LPM) Centroids}
\usage{
lpm.centroids(traj, id_field2 = FALSE, n_centroids = 3)
}
\arguments{
\item{traj}{A matrix or data.frame with each row representing the trajectory of observations of a unique location. The columns show the observation at consecutive time steps.}

\item{id_field2}{Whether the first column is a unique (id) field. Default: \code{FALSE}.}

\item{n_centroids}{Number of initial (linear) centroids to generate based on lpm technique}
}
\value{
A list of \code{n_centroids} linear trend lines
}
\description{
This function to create the initial centroids based on linear partitioning medoids (lpm) initialisation \code{(Adepeju et al. 2019, submitted)}
}
\examples{
traj <- gm.crime.sample1
print(traj)
traj <- missingValue(traj, id_field = TRUE, method = 2, replace_with = 1,
fill_zeros = FALSE) #filling the missing values
print(traj)
result_ <- lpm.centroids(traj, n_centroids=3, id_field = TRUE)
print(result_)
}
\references{
\code{Adepeju M, Langton S, Bannister J. (2019). akmeans: Anchored k-means: A longitudinal clustering technique for measuring long-term inequality in the exposure to crime at the micro-area levels (submitted)}.
}
