% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpm_centroids.R
\name{lpm_centroids}
\alias{lpm_centroids}
\title{Linear Partition Medoids (LPM) Centroids}
\usage{
lpm_centroids(dat, id_field2 = FALSE, n_centroids = 3)
}
\arguments{
\item{dat}{A matrix or data.frame with each row representing the trajectory of observations of a unique location. The columns show the observation at consecutive time steps.}

\item{id_field2}{Whether the first column is a unique (id) field. default: \code{FALSE}}

\item{n_centroids}{Number of initial (linear) centroids to generate based on lpm technique}
}
\value{
l_centroids
}
\description{
This function to create the initial centroids based on linear partitioning medoids (lpm) initialisation \code{(Adepeju et al. 2019, submitted)}
}
\references{
Adepeju M, Langton S, Bannister J. (2019). akmeans: Anchored k-means: A longitudinal clustering technique for measuring long-term inequality in the exposure to crime at the micro-area levels (submitted).
}
