% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingV_filler.R
\name{missingV_filler}
\alias{missingV_filler}
\title{Data imputing for longitudinal data}
\usage{
missingV_filler(traj, id_field = FALSE, method = 2, replace_with = 1, fill_zeros = FALSE)
}
\arguments{
\item{traj}{A matrix or data.frame with each row representing the trajectory of a unique location. The columns show the observations at consecutive time steps.}

\item{id_field}{Whether the first column is a unique (\code{id}) field. default: \code{FALSE}}

\item{method}{Method for calculating the missing values. Available options: \code{1}: arithmetic, \code{2}: regression. default: \code{1}}

\item{replace_with}{How to calculate the missing value. For \code{arithmetic} method: \code{replace_with} options are: \code{1}: Mean value of column, \code{2}: Minimum value of column, \code{3}: Maximum value of column, \code{4}: Mean value of row, \code{5}: Minimum value of row, or \code{6}: Maximum value of row. For \code{regression} method: the only available option for \code{replace_with} is: \code{1}: linear. That is, use a linear regression to interpolate or extrapolate the missing data values. Note: only the missing data points derive their new values from the regression line while the rest of the data points retain their original values. Trajectories with only one observation will be removed.}

\item{fill_zeros}{Whether to consider zeros (\code{0}) as missing values. default: \code{FALSE}. Only available for \code{2}: \code{regression} method.}
}
\value{
A data.frame with missing values \code{(NA, Inf, 0)} filled up
}
\description{
This function fills up any missing entries \code{(NA, Inf, 0)} in a matrix or dataframe using a value derived using a chosen method.
}
\details{
Given a matrix or data.frame with some missing values represented by \code{(NA, Inf, 0)}, the function \code{missingV_filler} determines the missing values using either the \code{arithmetic} or \code{regression} method.
}
\examples{
traj <- assault_data
print(traj)
missingV_filler(traj, id_field = TRUE, method = 2, replace_with = 1, fill_zeros = FALSE)
}
