% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_to_alpha.R
\name{num_to_alpha}
\alias{num_to_alpha}
\title{Function to assign alphabetic labels to numeric cluster IDs}
\usage{
num_to_alpha(x)
}
\arguments{
\item{x}{A vector of numeric cluster ids}
}
\description{
GGGGGGGGGGG
}
