% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_clust.R
\name{plot_clust}
\alias{plot_clust}
\title{To plot the clusters}
\usage{
plot_clust(data_clusters_list, id_field = TRUE)
}
\arguments{
\item{data_clusters_list}{A data.frame of clusters from \code{\link{akmeans_clust}}, in which the last column represents alphabetical cluster ids (labels)}

\item{id_field}{Whether the first column is a unique (id) field. [default: TRUE]}
}
\value{
data_clusters_list
}
\description{
To plot the clusters
}
