% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotr.R
\name{plotr}
\alias{plotr}
\title{Cluster plotting}
\usage{
plotr(clustr, id_field = TRUE, y.scaling = "fixed")
}
\arguments{
\item{clustr}{A data.frame of clusters from \code{\link{akmeans.clust}}, in which the last column represents alphabetical cluster ids (labels)}

\item{id_field}{Whether the first column is a unique (id) field. Default: \code{TRUE}.}

\item{y.scaling}{To set the vertical scales of the cluster panels. Options are: \code{"fixed"}: uses uniform scale for all panels, \code{"free"}: uses variable scales for panels.}
}
\value{
A plot showing clusters of trajectories, including the associated trendline - a linear trendline if clusters are based on \code{\link{lpm.centroids}} initialisation.
}
\description{
Function to plot the clusters generated by the \code{\link{akmeans.clust}} function
}
\details{
Given a matrix or data.frame with some suspected outlier observations, this function identified those observations based on the method chosen and replace all the observations accordingly.
}
\examples{
traj <- gm.crime.sample1
print(traj) #
traj <- missingValue(traj, id_field = TRUE, method = 2, replace_with = 1,
fill_zeros = FALSE) #filling up the missing values
print(traj)
clusters <- akmeans.clust(traj, id_field = TRUE, init_method = "lpm", n_clusters = 5)
dev.new()
print(plotr(clusters, id_field=TRUE, y.scaling="free"))
}
\references{
\code{Wickham H. (2016). Elegant graphics for Data Analysis. Spring-Verlag New York (2016)}
}
\keyword{clusters}
\keyword{plot,}
