% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpm.centroids.R
\name{qpm.centroids}
\alias{qpm.centroids}
\title{Quadratic Partition Medoids (QPM) Centroids}
\usage{
qpm.centroids(dat, n_centroids = 3, id_field = FALSE)
}
\arguments{
\item{dat}{A matrix or data.frame with each row representing the trajectory of observations of a unique location. The columns show the observation at consecutive time steps.}

\item{n_centroids}{Number of initial (quadratic) centroids to generate based on the qpm method (See attached Vignette)}

\item{id_field}{Whether the first column is a unique (id) field. Default: \code{FALSE}.}
}
\value{
q_centroids
}
\description{
Quadratic Partition Medoids (QPM) Centroids
}
