% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_all_whiteSpaces.R
\name{remove_all_whiteSpaces}
\alias{remove_all_whiteSpaces}
\title{Function to remove whitespace in data entries}
\usage{
remove_all_whiteSpaces(dat, head = TRUE)
}
\arguments{
\item{dat}{A matrix or data.frame}

\item{head}{If column names exist}
}
\value{
dat_Cleaned
}
\description{
Function to remove whitespace in data entries
}
