% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whiteSpaces.R
\name{whiteSpaces}
\alias{whiteSpaces}
\title{Removing whitespaces}
\usage{
whiteSpaces(traj, head = TRUE)
}
\arguments{
\item{traj}{A matrix or data.frame with each row representing the trajectory of observations of a unique location. The columns show the observation at consecutive time steps.}

\item{head}{Whether the field (column) names of the data.frame should be retained.Default: \code{TRUE}.}
}
\value{
A cleaned data.frame with whitespaces (if any) removed.
}
\description{
This function removes all leading and trailing whitespaces in longitudinal data
}
\references{
\url{https://en.wikipedia.org/wiki/Whitespace_character}
}
