% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akmeans_clust.R
\name{akmeans_clust}
\alias{akmeans_clust}
\title{akmeans_clust}
\usage{
akmeans_clust(dat, id_field = FALSE, init_centroids = "lpm",
  n_clusters = 3)
}
\arguments{
\item{dat}{A matrix or data.frame with each row representing the trajectory of observations of a unique location. The columns show the observation at consecutive time steps.}

\item{id_field}{Whether the first column is a unique (id) field. [default: FALSE]}

\item{init_centroids}{initialisation method [default: "lpm" - linear partitioning medoids @seealso \code{\link{lpm_centroids}}]}

\item{n_clusters}{number of clusters to generate [default (minimum value): 3]}
}
\value{
data_clusters_list
}
\description{
GGGGGGGGG
}
